pyproject.toml
[project]
name = "BossBot"
version = "0.1.0"
description = "Cutsomized LLM to aid with contract negotiation"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "chromadb",
    "fastapi",
    "fortune-python",
    "ipython",
    "uvicorn",
    "python-dotenv",
    "requests",
    "tqdm",
]
